+++ src/apis/user/info.py
@@ -581,6 +581,8 @@ async def post_tracker_switch(request: Request, response: Response, uid: Optiona
             tracker_in_use = 3
         elif tracker_in_use == "custom":
             tracker_in_use = 4
+        elif tracker_in_use == "unitracker":
+            tracker_in_use = 5
         else:
             response.status_code = 400
             return {"error": ml.tr(request, "invalid_value", var = {"key": "tracker"}, force_lang = au["language"])}