+++ src/apis/user/manage.py
@@ -41,6 +41,8 @@ async def post_accept(request: Request, response: Response, uid: int, authorizat
                 tracker_in_use = 3
             elif tracker_in_use == "custom":
                 tracker_in_use = 4
+            elif tracker_in_use == "unitracker":
+                tracker_in_use = 5
             else:
                 response.status_code = 400
                 return {"error": ml.tr(request, "invalid_value", var = {"key": "tracker"}, force_lang = au["language"])}
@@ -54,6 +56,8 @@ async def post_accept(request: Request, response: Response, uid: int, authorizat
                     tracker_in_use = 3
                 elif app.config.trackers[0]["type"] == "custom":
                     tracker_in_use = 4
+                elif app.config.trackers[0]["type"] == "unitracker":
+                    tracker_in_use = 5
             else:
                 response.status_code = 400
                 return {"error": ml.tr(request, "config_value_is_empty", var = {"item": "tracker"}, force_lang = au["language"])}