+++ src/config.py
@@ -1301,7 +1301,7 @@ def validateConfig(cfg):
             continue
         if "ip_whitelist" not in tracker.keys():
             tracker["ip_whitelist"] = []
-        if tracker["type"] not in ["tracksim", "trucky", "custom"]:
+        if tracker["type"] not in ["tracksim", "trucky", "custom", "unitracker"]:
             continue
         if tracker["ip_whitelist"] is not None and type(tracker["ip_whitelist"]) != list:
             continue